(feat) added: relation between user and role, and estructuring logger as a service
