+++ src/modules/security/role/entities/role.entity.ts
@@ -1,1 +1,14 @@
-export class Role {}
+import { Entity, PrimaryGeneratedColumn, Column, OneToMany } from 'typeorm';
+import { User } from '../../user/entities/user.entity';
+
+@Entity({ schema: 'security' })
+export class Role {
+  @PrimaryGeneratedColumn()
+  id: number;
+
+  @Column()
+  name: string;
+
+  @OneToMany(() => User, user => user.role)
+  users: User[];
+}
