+++ src/modules/security/role/role.module.ts
@@ -1,9 +1,12 @@
 import { Module } from '@nestjs/common';
 import { RoleService } from './role.service';
 import { RoleController } from './role.controller';
+import { Role } from './entities/role.entity';
+import { TypeOrmModule } from '@nestjs/typeorm';
 
 @Module({
   controllers: [RoleController],
   providers: [RoleService],
+  imports: [TypeOrmModule.forFeature([Role])],
 })
 export class RoleModule {}
