+++ src/services/logger/enums/LogColors.enum.ts
@@ -1,0 +1,10 @@
+export enum LogColorsEnum {
+    red = '\x1b[31m',
+    green = '\x1b[32m',
+    yellow = '\x1b[33m',
+    blue = '\x1b[34m',
+    magenta = '\x1b[35m',
+    cyan = '\x1b[36m',
+    pink = '\x1b[38;5;206m',
+    close = '\x1b[0m',
+  }
