+++ src/services/logger/enums/LogLevel.enum.ts
@@ -1,0 +1,8 @@
+export enum LogLevelEnum {
+    ERROR = 'error',
+    WARNING = 'warning',
+    INFO = 'info',
+    DEBUG = 'debug',
+    FATAL = 'fatal',
+    TRACE = 'trace'
+  }
