+++ src/services/logger/interfaces/LogMetadata.interface.ts
@@ -1,0 +1,7 @@
+export interface LogMetadata {
+  clientIp: string;
+  controller: string;
+  method: string;
+  url: string;
+  miliseconds: string;
+}
